Refactor : Refactor user review, manner serializer
